+++ tests/unit/Util/FileMatcherTest.php
@@ -612,7 +612,7 @@ class FileMatcherTest extends TestCase
     {
         $this->expectException(RuntimeException::class);
         $this->expectExceptionMessage('Path "foo/bar" must be absolute');
-        FileMatcher::match('foo/bar', new FileMatcherPattern(''));
+        FileMatcher::toRegEx('/a')->matches('foo/bar');
     }
 
     /**
@@ -621,7 +621,7 @@ class FileMatcherTest extends TestCase
     private static function assertMap(FileMatcherPattern $pattern, array $matchMap): void
     {
         foreach ($matchMap as $candidate => $shouldMatch) {
-            $matches = FileMatcher::match($candidate, $pattern);
+            $matches = FileMatcher::toRegEx($pattern->path)->matches($candidate);
 
             if ($matches === $shouldMatch) {
                 self::assertTrue(true);
